+++ src/sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added license information in the generated Nuget packages #945 [clurdish - Landon Petzoldt]
 - `Author` property for `IWeb` #947 [jansenbe - Bert Jansen]
 - Static method to create a PnPContext mock for use in testing #948 [jansenbe - Bert Jansen]
+- Support for getting, setting and clearing of (folder based) column defaults #949 [jansenbe - Bert Jansen]
 
 ### Changed
 
@@ -24,6 +25,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Add KANBAN to the ViewType2 enum #930 [martinlingstuyl - Martin Lingstuyl]
 - Add MODERNCALENDAR  to the ViewType2 enum #939 [martinlingstuyl - Martin Lingstuyl]
 - Fix search sort direction with new POST approach #937 [lroos - Llewellyn Roos]
+- Ensure the Name property is loaded #955 [jansenbe - Bert Jansen]
 
 ## [1.7.0]